refactor(elements): Move components to single file
