+++ packages/elements/src/index.tsx
@@ -6,11 +6,7 @@ export { Form, Input, Field, Label, Submit } from './common/form';
 export { SocialProviders } from './common/social-providers';
 
 /** Sign In Components */
-export { SignIn } from './sign-in/sign-in';
-export { SignInFactorOne } from './sign-in/factor-one';
-export { SignInFactorTwo } from './sign-in/factor-two';
-export { SignInStart } from './sign-in/start';
-export { SignInSSOCallback } from './sign-in/sso-callback';
+export { SignIn, SignInStart, SignInFactorOne, SignInFactorTwo, SignInSSOCallback } from './sign-in';
 
 /** Hooks */
 export { useSignInFlow, useSignInFlowSelector } from './internals/machines/sign-in.context';
