--- packages/elements/src/sign-in/factor-one.tsx
@@ -1,10 +1,0 @@
-import { Route } from '../internals/router-react';
-
-export function SignInFactorOne({ children }: { children: React.ReactNode }) {
-  return (
-    <Route path='factor-one'>
-      <h1>Factor One</h1>
-      {children}
-    </Route>
-  );
-}
